requirements.txt
# Core dependencies for YOLOv11
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.6.0
pillow>=9.0.0
numpy>=1.23.0

# FastAPI and server dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Additional utilities
matplotlib>=3.7.0
pyyaml>=6.0
requests>=2.31.0
